Fix: Create python files improperly on first run
Also reverted back to `O2` optimization because we dont need debugging anymore. Doing that will only case the speed to decrease.
On the other hand, properly converted `userByRowOrder.size()` from `size_t` to `int` with `static_cast`.

+++ main/WIN_UsersSettings.cpp
@@ -228,7 +228,7 @@ void WIN_UsersSettings::reloadVars() {
 }
 
 void WIN_UsersSettings::selectUserFromRow(int row) {
-    if (row < 0 || row >= userByRowOrder.size()) {
+    if (row < 0 || row >= static_cast<int>(userByRowOrder.size())) {
         std::cerr << "Invalid row: " << row << ", size: " << userByRowOrder.size() << '\n';
         return;
     }
